requirements.txt
# Core Data Processing
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Data Loading & Parsing
pyarrow>=12.0.0  # For Parquet format
openpyxl>=3.1.0  # Excel support

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Time Series Analysis
statsmodels>=0.14.0
pmdarima>=2.0.3  # Auto ARIMA
prophet>=1.1.4   # Facebook Prophet

# Machine Learning
scikit-learn>=1.3.0
xgboost>=1.7.0
lightgbm>=4.0.0

# Deep Learning (Optional - for LSTM models)
tensorflow>=2.13.0  # or pytorch
keras>=2.13.0

# Climate Data Specific
netCDF4>=1.6.0  # For climate data formats
xarray>=2023.1.0
metpy>=1.5.0  # Meteorological calculations

# Progress Bars & Utilities
tqdm>=4.65.0
jupyter>=1.0.0
ipykernel>=6.25.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Testing (Optional)
pytest>=7.4.0
